backend_python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
pydantic-settings==2.2.1
httpx==0.27.0
sse-starlette==1.8.2
python-multipart==0.0.9
Pillow==10.4.0
